Renamed dqn folder, fixed broken imports, updated docstring
